Add carnivore subproject
This game is going to be a coding challenge game, which will need a separate
distribution for the implementation details, that will also be included in the
game distribution.

I think the easiest way to implement this is with a subproject that contains
the code that needs to be shared, which can then be directly included in the
game through a project dependency, whilst the subproject can be published and
distributed separately for the players.

This will save on duplication if we used two separate projects, and will
ensure code compatibility as long as the same version of game and carnivore is
used

As for the implementation- it's just an interface right now which has
functions for initialisation, updating, and bidding ready to be implemented.
These should match the functionality I expect in the game, but may require
tweaks after hooking things up.