Configure spotless with diktat
Diktat is a bit stricter with it's formatting rules, and enforces some stuff
that you'd expect at static analysis level. Catching this early is better
than late, so I've included this in the spotless step so the code quality is
in theory, better.

So far have only had to add KDoc where we didn't care before. This is actually
a good change as the KDoc is going to help me remember in the future, and it
has forced me to write some instructions for the distributable player
interface which is better than not having it, and I might have forgotten
otherwise.

May tweak the rules as time goes on, but we'll see what happens.