Add background to title screen
First we're setting up git lfs, so we're not pushing images up to our git
repository. I think this works with GitHub out of the box, I can't see
anything that says otherwise when I looked. I'll track additional files as
needed and hopefully not forget.

Replaces the badlogic logo with an actual background. This is placeholder
art that has been borrowed from a Google search that I can't find anymore, but
will be replaced at some point so no licensing issues. In the long term
anyway.