Add title image to Title Screen
Couple of changes here. First, adding more ktx dependencies for scene2d that
I missed before. This one is the actors dependency, which allows us to center
the title with reduced maths required.

Also adds the title image. This is just text currently, but is supplied as an
image. This is because I'd like to make it less basic in the future without
having to change a lot, and actually more importantly, it's a lot easier to
outline the text when it's an image rather than actual text.

The title uses the Dinopia font by Little Red Studio. This is licensed through
Creative Fabrica (ID 33929772) which includes commercial use, but I'm not sure
if I'm going to need that. Better safe than sorry though.

I really like this Dinopia font by the way, plan to use it everywhere because
it's just so good.