Turn off magic number check
This is probably not a good idea, but I'm not convinced defining a meaningful
variable for every width and offset is a great idea. Maybe it is? Like I could
define `meatOffset` or something instead of just hardcoding 25, but I don't
know

Let's see if I come to regret this.