Add some shaky meat
This adds a piece of meat that rotates back and forth in a little jig for a
little bit of excitement on the title screen.

I've set the image size to be 256 x 256, so it fits nicely where it needs to
go.

The meat is from flaticon, designed by Freepik, and is licensed with the
premium subscription, so no attribution required for use.

It might be a placeholder, not sure if I will bother changing it up in the
future or not.