Adds fading instructions
This adds some "Press Space to Start" instructions to the title screen, which
fade in and out continuously, so you know the game is running and not crash,
as if the meat dance wasn't enough.

Uses Dinopia font again because it's great. This is also an image rather than
text, this might be able to be a Scene2D label or something like that in the
future.