Add a little dinosaur rainbow
This adds a little rainbow of carnivores to the bottom of the title screen,
one for each colour that we're going to make available for players to choose
from.

To save duplication, this adds the colours to an enum in the carnivore module,
which can be used by players to choose the colour of their carnivore. We can
then iterate the values in that enum to display one of each. By luck or by
some very careful planning, there's just enough room to show one of each
colour on the title screen.

The carnivore is a stock image again, this time from Adobe Stock, it's called
"Cute dinosaur-baby.Small green dinosaur who just hatched from the egg",
designed by Oleg Lytvynenko. ID is 199615704. Licensed using the standard
license from Adobe.

Possibly a placeholder, but it was a reasonable amount of work to recolour
them all, so I might just keep them. We'll see.