Hats
Everyone loves hats. This adds the capability for the carnivores to wear hats.

The only hat right now is the cowboy hat, but I'll add a lot more before
actually working on getting the game working, because I am a ridiculous human.

This will eventually shuffle all the hats and then pop them off a list, so the
hats on the title screen will be different each time, without the possibility
of duplicates.

I might include the hats and carnivore sprites in the carnivore project, as I
think they will still get built into the jar, and it would allow the carnivore
distributable to contain the sprites, so players will be able to make a more
informed choice.

I'll test it out next, so the next commit will either be moving the sprites,
or it'll be a new hat because it didn't work.