Move the carnivore sprites
The sprites that are under the control of the player have been moved to the
carnivore project, so they can see them in that project when making the choice
of colour and hat.

I'll be able to add them to the README easily this way too, which will be neat.