Add a game screen
An actual game screen that will contain actual gameplay, after all this time.

So far this updates the game to add the new game screen, adds a click listener
to the title screen that is listening for the space key, and swaps to the game
screen when space is pressed.

I needed to change the stage from being a singleton to being provided by each
inject call, as all the actors from the title screen were preserved in the
game screen. I don't think that's what I want to happen, so using an entirely
new stage seems like a reasonable thing to do.