Change music to play throughout game
This moves the music playing from the title screen to the main game class, as
we want the same music to play throughout the game even when starting the
actual game itself.

Also changes the music to set up a completion event listener that will cause
the music to play again, making it repeat.