Add meat content to game screen
These commit messages are going to be funky when reading them back, like what
in the world is "meat content"?

A lot of changes again for what seems like a simple enough thing to do, with
the meat content.

Adds some sprites to the game screen. Background and meat are very similar to
the title screen.

The interesting part is that we want to display the amount of meat up for
grabs by the carnivores in the game screen, so we need to know how much there
is.

Thematically, I've introduced a `MummyTrex` class that handles the amount of
meat via `bringHomeTheBacon`, because it's funny. Kind of. When you create a
new `MummyTrex` she initialises the amount of meat available using a random
value between 1000 and 2000. This is then exposed to the game. We'll have to
introduce a way to modify the amount of meat available after a successful bid,
so this might not be immutable in the future, but it'll do for now.

To show this to the people playing the game, we're adding a label to the game
screen, semi-centered on the meat itself. This label uses the Dinopia font
we used before for the title and instructions, but this time as actual text
instead of an image. This means we need to include the font and create it for
use within our game. As part of this I found out you can set outlines
programmatically. I might go back and use this instead of images in the title
screen, but I'm not 100% convinced it looks as good. We'll see.

What else has happened. I added a couple of extra ktx libraries for FreeType
and Skins, which make it easier to define a skin with a font for the label.

I think that's about it actually. Next I need a way to read the players from
the directory, so we can display them at the bottom of the game screen. This
part should be fun.