Adds a way to load players
This adds a CarnivoreLoader, which is used to load the jars in the carnivores
directory on the local file system.

We iterate through the jars and then load all the classes present in the jars,
if we find a class that implements the `CraftyCodeCarnivore` while we're
loading, we return add it to the list and move on to the next jar.

If carnivores can't be found, we'll load the FairCarnivore that we currently
include in the distributed game. It might not be in the future, we'll see how
it goes.