Add carnivore loading functionality
This hooks the previous carnivore loader into the game screen, where we use it
to load the carnivores from the file system and display them on the game
screen.

After loading the carnivores, we loop over them and use their colour and hat
declarations to display their avatar on the screen.