Make the fair carnivore more random
This sets a random colour (no duplicates), random hat (no duplicates), and a
more random bid amount (possible duplicates) to the fair carnivore to make it
a bit more interesting to watch.

I think I will be updating the carnivore loader to fill out remaining spaces
when there are not enough players with fair carnivores, so having them behave
slightly differently is a good thing.