Add CarnivoreActor
The carnivore actor is an orchestration piece that will be used as a container
for all the other actors that are related to that carnivore. This will allow
us to tell the carnivore what to do on a high level, like "show bid" and have
the actual logic for how that happens live somewhere else.

I don't know if it's a great idea yet. I want to separate the game logic from
LibGDX and Scene2D, so I can run simulations without a UI, but I haven't done
a very good job of that yet. Maybe I should have separated the game logic
from the start and then wove it into the UI.

Oh well, it'll be ok.