Add game logic
This adds the rest of the game logic to play each round.

Rounds are controlled by hitting any key, which is enabled when all animations
have been played for a round.

After initialising all carnivores with the amount of meat being played for,
each round we ask each carnivore to bid for the amount of meat. We select the
carnivore with the lowest amount of meat bid, and they are the "winner" for
that round and are removed from bidding on future rounds.

After the winning bid is made, the remaining carnivores are told the amount of
meat that has been eaten, and mummy trex is updated with the same information.

Rounds proceed that way until every carnivore has gotten some meat, after
which we could crown a winner.

We don't though, proceeding to the next round crashes the game. At least until
I put in the victory screen, at which point the game will be... complete?

I've never completed anything before. This is exciting.