Add publishing configuration
This adds all the publishing configuration required to stage, close, and
release a repository in Sonatype, which will allow us to publish our artifacts
to Maven Central.

I'm not massively bothered about publishing the game itself to Central,
although it will be nice to have it available somewhere that I don't have to
pay for bandwidth.

Distributing the `carnivore` artifact is the main purpose here, so
players are able to pull it into their projects and build something compatible
with the game.

I've tested this locally and everything works, will test it with GitHub
actions next to make sure I don't have to do anything manually from now on