Add publish workflow
This should pull all the required secrets from GitHub and upload artifacts to
the Sonatype staging repository.

This is the test commit testing against this branch to make sure it all works,
once it does I'll update this to perform a full release and trigger it from
the main branch.