Add an instructions screen in game
This contains the content from the README inside the game itself. This might
not be massively useful, but it felt like it needed to be included for
completeness, and maybe a little bit of procrastination.

You can't get to this page yet, but now that I know how to do buttons, that
will be a piece of cake.