Add buttons to title screen
This adds a "Play", an "Instructions", and a "Quit" button to the title
screen which can be used to navigate to different parts of the game. It looks
very cool, if I do say so myself. And was so easy!

Removes the "Press space to play" instructions from before, and deletes the
asset required for that. We won't be needing it anymore. Also removes the
ability to press space to play. I will need to add a way to navigate the
buttons using a keyboard at some point, but not today.

Reduces the amount of carnivores on the title screen to 6, to make room for
the buttons. It's a worthwhile sacrifice.