Update carnivore colours
Removes light blue as it made the selection an odd number which made me feel
weird inside.

Renames purple and pink to indigo and violet. They're... kind of indigo and
violet. I hope everyone knows the rainbow as little as I do so they don't
realise it's not that great.