Rename gradle subprojects
This changes the gradle subprojects to use directories with a single word, and
then maps them to the full names when setting up the projects.

They need the long names to be put in the right place in Sonatype. I think.
I thought the jar name would work, but it didn't, and I forgot to check. This
almost definitely works, so it should be ok.