Re order screens
This changes the order of the screens and the transitions between them. We now
go Title -> Game -> Simulation -> Victory, which makes more sense than having
a little carnivore celebrating before getting creamed in the simulations.

We don't have a way to crown the overall winner in the victory screen yet,
so currently the winner is whoever won last. Which is less than ideal.

To be continued.