Run more simulations
Before we were running one simulation per frame, so roughly 60 simulations per
second. To run 10,000 simulations it took around 3 minutes to complete.

We now run 20 simulations per frame, so about 1200 per second. That made the
simulation end too quickly, but bumping simulations to 100,000 means that we
finish running simulations in about 80 seconds, or roughly the amount of time
the game music lasts for. How's that for intelligent engineering?

The bigger numbers are also a lot cooler than they were before.