Resolve ties pseudorandomly
The behaviour we had before would resolve a tie by picking the first person in
the list and assigning them the win.

I don't know if this would cause a problem particularly, as a tie is unlikely
and not very interesting. But to resolve any potential issue, this tie is now
resolved pseudorandomly instead, similarly to the way it is resolved in the
simulation.