Implement conservation of mass
We were allowing carnivores to create as much meat as they want just by
thinking it, where they could win a bid with more meat than the amount that is
actually left.

This updates the Game to show the naughty bid in red, which then becomes a bid
of zero when evaluated. I don't know why I bothered with this as the game
screen isn't used all that often, but I don't know I think it's kind of cool,
even if it will never actually be seen. The code isn't great, but I want to
clean up the logic in general anyway.

The simulation screen has two fixes- one, you can't bid more than the amount
of meat remaining, like the game screen. Second, we actually play the game
now properly. Before, the amount of meat remaining was never updated in the
match, which would cause obviously weird results. The meat is now updated.