Fill up roster
This will fill up the roster with fair carnivores when there aren't 6 players,
which might not be something I end up keeping because what if someone wants to
play with 3 players or something? It'll have to be configurable, like adding
CPUs to a game.

Anyway this will do for now.